App/SICD6mA_predictor/requirements.txt
anaconda 5.3;
python 3.7;
pytorch 1.0;
numpy;
wx;
sklearn;
pylab;
os;


